Webhook manual trigger (#11687)
* Add webhookManualTrigger mutation.

Get rid of ROOT_TYPE mapping, extend WebhookTriggerErrorCodes, extend validation.

Cover with tests.

Capture event delivery status.

* Update versions and changelog.

* Delete eventDelivery of successfully delivered webhooks, return delivery=null for scuh cases.

* Refactor get_event_type_from_subscription function to return list of evetual multiple events.

* Allow inheriting events from query field for webhookUpdate and webhookCreate mutations.

* Fix test.

* enhance error handling in get_event_type_from_subscription function

* Add webhookManualTrigger mutation.

Get rid of ROOT_TYPE mapping, extend WebhookTriggerErrorCodes, extend validation.

Cover with tests.

Capture event delivery status.

* Update versions and changelog.

* Delete eventDelivery of successfully delivered webhooks, return delivery=null for such cases.

* Return delivery for successfully delivered webhook.

* Split validation into chunks.

* Webhook app clean up. (#11756)
