+++ CHANGELOG.md
@@ -11,8 +11,10 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add `webhookDryRun` mutation - #11548 by @zedzior
 - Fix adding invalid label to meta fields - #11718 by @IKarbowiak
 - Add filter by `checkoutToken` to `Query.orders`. - #11689 by @kadewu
+- Add `WebhookTrigger` mutation - #11687 by @zedzior
 
 ### Other changes
+- Allow `webhookCreate` and `webhookUpdate` mutations to inherit events from `query` field - #11736 by @zedzior
 
 # 3.10.0 [Unreleased]
