+++ saleor/graphql/webhook/mutations/webhook_update.py
@@ -1,0 +1,115 @@
+import graphene
+from django.core.exceptions import ValidationError
+
+from ....permission.enums import AppPermission
+from ....webhook import models
+from ....webhook.error_codes import WebhookErrorCode
+from ...core import ResolveInfo
+from ...core.descriptions import ADDED_IN_32, DEPRECATED_IN_3X_INPUT, PREVIEW_FEATURE
+from ...core.mutations import ModelMutation
+from ...core.types import NonNullList, WebhookError
+from .. import enums
+from ..subscription_payload import validate_query
+from ..types import Webhook
+from .webhook_create import clean_webhook_events
+
+
+class WebhookUpdateInput(graphene.InputObjectType):
+    name = graphene.String(description="The new name of the webhook.", required=False)
+    target_url = graphene.String(
+        description="The url to receive the payload.", required=False
+    )
+    events = NonNullList(
+        enums.WebhookEventTypeEnum,
+        description=(
+            f"The events that webhook wants to subscribe. {DEPRECATED_IN_3X_INPUT} "
+            "Use `asyncEvents` or `syncEvents` instead."
+        ),
+        required=False,
+    )
+    async_events = NonNullList(
+        enums.WebhookEventTypeAsyncEnum,
+        description="The asynchronous events that webhook wants to subscribe.",
+        required=False,
+    )
+    sync_events = NonNullList(
+        enums.WebhookEventTypeSyncEnum,
+        description="The synchronous events that webhook wants to subscribe.",
+        required=False,
+    )
+    app = graphene.ID(
+        required=False,
+        description="ID of the app to which webhook belongs.",
+    )
+    is_active = graphene.Boolean(
+        description="Determine if webhook will be set active or not.", required=False
+    )
+    secret_key = graphene.String(
+        description="Use to create a hash signature with each payload."
+        f"{DEPRECATED_IN_3X_INPUT} As of Saleor 3.5, webhook payloads default to "
+        "signing using a verifiable JWS.",
+        required=False,
+    )
+    query = graphene.String(
+        description="Subscription query used to define a webhook payload."
+        + ADDED_IN_32
+        + PREVIEW_FEATURE,
+        required=False,
+    )
+
+
+class WebhookUpdate(ModelMutation):
+    class Arguments:
+        id = graphene.ID(required=True, description="ID of a webhook to update.")
+        input = WebhookUpdateInput(
+            description="Fields required to update a webhook.", required=True
+        )
+
+    class Meta:
+        description = "Updates a webhook subscription."
+        model = models.Webhook
+        object_type = Webhook
+        permissions = (AppPermission.MANAGE_APPS,)
+        error_type_class = WebhookError
+        error_type_field = "webhook_errors"
+
+    @classmethod
+    def clean_input(cls, info, instance, data):
+        cleaned_data = super().clean_input(info, instance, data)
+        app = cleaned_data.get("app")
+
+        if not instance.app_id and not app:
+            raise ValidationError(
+                "Missing token or app", code=WebhookErrorCode.INVALID.value
+            )
+
+        if instance.app_id:
+            # Let's skip app id in case when context has
+            # app instance
+            app = instance.app
+            cleaned_data.pop("app", None)
+
+        if not app or not app.is_active:
+            raise ValidationError(
+                "App doesn't exist or is disabled",
+                code=WebhookErrorCode.NOT_FOUND.value,
+            )
+        clean_webhook_events(info, instance, cleaned_data)
+
+        if query := cleaned_data.get("query"):
+            validate_query(query)
+            instance.subscription_query = query
+        return cleaned_data
+
+    @classmethod
+    def save(cls, _info: ResolveInfo, instance, cleaned_input):
+        instance.save()
+        events = set(cleaned_input.get("events", []))
+        if events:
+            instance.events.all().delete()
+            models.WebhookEvent.objects.bulk_create(
+                [
+                    models.WebhookEvent(webhook=instance, event_type=event)
+                    for event in events
+                ]
+            )
