+++ saleor/plugins/webhook/tasks.py
@@ -499,6 +499,7 @@ def send_webhook_request_async(self, event_delivery_id):
                 delivery.event_type,
                 data,
             )
+
         attempt_update(attempt, response)
         if response.status == EventDeliveryStatus.FAILED:
             task_logger.info(
